BN-4302 [FEATURE] Rename base module to NgNetworkModuleCore
