+++ Sources/NgNetworkModuleCore/Common/Extensions/Data+Codable.swift
