+++ Sources/NgNetworkModuleCore/NetworkError/NetworkError.swift
