+++ Sources/NgNetworkModuleCore/NetworkError/NetworkErrorConvertible.swift
