+++ Sources/NgNetworkModuleCore/NetworkModule/Helpers/AuthenticationTokenProvider.swift
