+++ Sources/NgNetworkModuleCore/NetworkResponse/NetworkResponse.swift
