+++ Tests/NgNetworkModuleCoreTests/NetworkError/NetworkErrorConvertibleTests.swift
@@ -6,7 +6,7 @@
 import Foundation
 import XCTest
 
-@testable import NgNetworkModule
+@testable import NgNetworkModuleCore
 
 final class NetworkErrorConvertibleTest: XCTestCase {
     var sut: NetworkErrorConvertible!
