+++ Tests/NgNetworkModuleCoreTests/NetworkModule/Helpers/AuthenticationTokenProviderTests.swift
@@ -6,7 +6,7 @@
 import Foundation
 import XCTest
 
-@testable import NgNetworkModule
+@testable import NgNetworkModuleCore
 
 final class AuthenticationTokenProviderTest: XCTestCase {
     var sut: AuthenticationTokenProvider!
